cv/mobileclip_app/model_conversion/requirements.txt
# PyTorch to TFLite/ONNX conversion tools
onnx>=1.15.0
onnxruntime>=1.16.0
tf2onnx>=1.16.0

# TensorFlow for TFLite conversion and testing
tensorflow>=2.15.0

# Image processing
pillow>=10.0.0
opencv-python>=4.8.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Utils
numpy>=1.24.0
tqdm>=4.66.0

# MobileCLIP dependencies (should be already installed)
# torch>=2.0.0
# torchvision>=0.15.0
